+++ pollux/lib/sql-doobie/src/main/resources/sql/pollux/V4__add_thid_unique_constraint.sql
@@ -1,2 +1,0 @@
-ALTER TABLE public.issue_credential_records
-ADD CONSTRAINT unique_thid UNIQUE (thid);